pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "distoptica"
dynamic = ["version"]
description = """
Python library for modelling optical distortions.\
"""
readme = "README.md"
authors = [
    {name = "Matthew Fitzpatrick", email = "matthew.rc.fitzpatrick@gmail.com"}
]
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Natural Language :: English"
]
keywords = [
    "pytorch",
    "microscopy",
    "distortion",
    "distortion correction"
]
dependencies = [
    "fancytypes>=1.0.0",
    "torch"
]
requires-python = ">=3.8"

[project.optional-dependencies]
tests = [
    "pytest-cov"
]
examples = [
    "jupyter",
    "ipympl",
    "hyperspy",
    "numba"
]
docs = [
    "sphinx<7",
    "sphinx_rtd_theme",
    "sphinx-autodoc-typehints",
    "numpydoc",
    "docutils"
]
all = [
    "distoptica[tests, examples, docs]"
]

[project.urls]
Homepage = "https://mrfitzpa.github.io/distoptica"
Repository = "https://github.com/mrfitzpa/distoptica"

[tool.setuptools_scm]
write_to = "distoptica/version.py"
version_scheme = "post-release"